+++ src/atoms/effortValueAtom.ts
@@ -1,6 +1,6 @@
 import { atom } from "jotai";
 
-import { ATK, DEF, HP, SP_ATK, SP_DEF, SPD } from "@/features/constants";
+import { ATK, DEF, HP, SP_ATK, SP_DEF, SPD } from "@/constants";
 import type { PokemonStatus } from "@/types";
 
 export const initialState = [
