+++ src/constants/term.ts
@@ -1,0 +1,8 @@
+export const HP = "hp" as const;
+export const ATK = "attack" as const;
+export const DEF = "defense" as const;
+export const SP_ATK = "special-attack" as const;
+export const SP_DEF = "special-defense" as const;
+export const SPD = "speed" as const;
+
+export const STATUS_SPECIES = [HP, ATK, DEF, SP_ATK, SP_DEF, SPD] as const;
