+++ src/features/status/components/StatusTableBody/presentation.tsx
@@ -6,8 +6,8 @@ import {
 
 import type { StatusSpecies } from "@/types";
 
-import { MAX_TOTAL_EFFORT_VALUE, MAX_EFFORT_VALUE, MAX_INDIVIDUAL_VALUE } from "../../constants";
 import { toJaStatusSpecies } from "../../logic/toJaStatusSpecies";
+import { MAX_TOTAL_EFFORT_VALUE, MAX_EFFORT_VALUE, MAX_INDIVIDUAL_VALUE } from "../styleConfig";
 import { InputFieldWithButton, InputField } from "../table/TableRow/ui";
 
 import type { MouseEventHandler } from "react";
