+++ src/features/status/components/styleConfig.ts
