+++ src/features/status/components/table/TableRow/ui/InputFieldWithButton.tsx
@@ -5,7 +5,7 @@ import {
   VStack,
 } from "@chakra-ui/react";
 
-import { INPUT_GROUP_WIDTH } from "@/features/status/constants";
+import { INPUT_GROUP_WIDTH } from "@/features/status/components/styleConfig";
 
 import { InputField } from "../ui";
