+++ src/features/status/logic/term.ts
@@ -6,12 +6,3 @@ export const SP_DEF_JA = "とくぼう" as const;
 export const SPD_JA = "すばやさ" as const;
 
 export const STATUS_SPECIES_JA = [HP_JA, ATK_JA, DEF_JA, SP_ATK_JA, SP_DEF_JA, SPD_JA] as const;
-
-export const HP = "hp" as const;
-export const ATK = "attack" as const;
-export const DEF = "defense" as const;
-export const SP_ATK = "special-attack" as const;
-export const SP_DEF = "special-defense" as const;
-export const SPD = "speed" as const;
-
-export const STATUS_SPECIES = [HP, ATK, DEF, SP_ATK, SP_DEF, SPD] as const;
