+++ src/mocks/mockData/baseStatsMockData.ts
@@ -1,4 +1,4 @@
-import { ATK, DEF, HP, SP_ATK, SP_DEF, SPD } from "@/features/constants";
+import { ATK, DEF, HP, SP_ATK, SP_DEF, SPD } from "@/constants";
 import type { PokemonStatus } from "@/features/types";
 
 type MockData = {
